Create helloworld.java
Write Java code for the below exercises and then check your code against the solutions below.
Exercise 1. Declare a variable to store the literal value "Jon Smith", assign it that value, and print the value of the variable to the console.
Output example:
Jon Smith
Exercise 2. Declare a variable to store book title, another to store the number of pages, and last one to store the book's price. Assign the following values to the variables: "Black Dawn", 567, and 9.99. Print the values of the three variables to the console on the same line with meaningful labels.
Output example:
Book Title="Black Dawn" Number of pages=567 Price=$9.99
Exercise 3. Declare the following variables: variable to store whether an account earns an interest or not, variable to store account's id, variable to store account owner's name, and variable to store account's balance. Assign valid values to all variables. Print the information to the console across multiple lines with meaningful labels.

Output example:
Account information:
   owner=Jane Jones id=56789
   balance=$567.45 earn interest: true